chore(processor): move to helper
