+++ processor/src/services/converters/create-payment.converter.ts
@@ -3,7 +3,7 @@ import { config } from '../../config/config';
 import { ThreeDSRequestData } from '@adyen/api-library/lib/src/typings/checkout/threeDSRequestData';
 import { Cart, CurrencyConverters, Payment } from '@commercetools/connect-payments-sdk';
 import { buildReturnUrl, getShopperStatement, populateApplicationInfo, populateCartAddress } from './helper.converter';
-import { mapCoCoCartItemsToAdyenLineItems } from './lineitem.converter';
+import { mapCoCoCartItemsToAdyenLineItems } from './helper.converter';
 import { CreatePaymentRequestDTO } from '../../dtos/adyen-payment.dto';
 import { getFutureOrderNumberFromContext } from '../../libs/fastify/context/context';
 import { paymentSDK } from '../../payment-sdk';
